refactor(SIA::GUI): add Similar Item work view
