+++ src/ramstk/views/gtk3/allocation/__init__.py
@@ -1,11 +1,11 @@
 # pylint: disable=unused-import
 # -*- coding: utf-8 -*-
 #
-#       ramstk.views.gtk3.function.__init__.py is part of The RAMSTK Project
+#       ramstk.views.gtk3.allocation.__init__.py is part of The RAMSTK Project
 #
 # All rights reserved.
-# Copyright 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
-"""The RAMSTK GTK3 revision package."""
+# Copyright 2007-2020 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
+"""The RAMSTK GTK3 allocation package."""
 
 # RAMSTK Local Imports
 from .workview import Allocation as wvwAllocation
