+++ src/ramstk/views/gtk3/similar_item/__init__.py
@@ -1,0 +1,12 @@
+# pylint: disable=unused-import
+# -*- coding: utf-8 -*-
+#
+#       ramstk.views.gtk3.similar_item.__init__.py is part of The RAMSTK
+#       Project
+#
+# All rights reserved.
+# Copyright 2007 - 2020 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
+"""The RAMSTK GTK3 similar item package."""
+
+# RAMSTK Local Imports
+from .workview import SimilarItem as wvwSimilarItem
